Update context_processors.py
there are no buttons in the database currently, and i had some in the past, and  i ended up calling them in the templates, very convenient, as long as the data is in the db, im gonna need to rethink this one. but for now a simple check for None and returning a default button should work as a decent default for now. resolves #8